# back/openapi.yaml
openapi: 3.0.3
info:
  title: IS Course Work Back API
  description: >-
    REST API that powers the restaurant project. It exposes
    endpoints to manage employees, dishes, ingredients, tables, orders, bills,
    feedback, reports, and employee wallets. Most operations require a JWT that
    encodes the employee id and position (role).
  version: 1.0.0
servers:
  - url: http://localhost:8080/is_coursework/api
    description: Local application server
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new employee
      description: >-
        Creates an employee together with an associated wallet and returns a
        one-time numeric code that can be used to log in for the first time.
        Only administrators may register new employees.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Employee'
      responses:
        '201':
          description: Employee registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthData'
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in with an employee id and code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthData'
      responses:
        '200':
          description: JWT issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
  /bills/create:
    post:
      tags: [Bills]
      summary: Create a bill for a table
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BillCreationRequest'
      responses:
        '201':
          description: Bill created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillCreationResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/ServerError' }
  /bills/{id}:
    get:
      tags: [Bills]
      summary: View bill
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: birthday
          in: query
          description: Whether the guest has a birthday (adds bonus points)
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Bill info with bonus points
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillLookupResponse'
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
  /bills/pay/{id}:
    post:
      tags: [Bills]
      summary: Set status paid for a bill
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Bill paid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BillPaymentResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '500': { $ref: '#/components/responses/ServerError' }
  /dish/get-all:
    get:
      tags: [Dishes]
      summary: Get all dishes
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Collection of dishes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Dish'
        '500': { $ref: '#/components/responses/ServerError' }
  /dish/create:
    post:
      tags: [Dishes]
      summary: Create dish
      description: Only administrators may create dishes.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DishWithIngredientsRequest'
      responses:
        '200':
          description: Dish created
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /dish/modify-dish:
    post:
      tags: [Dishes]
      summary: Update dish properties
      description: Only administrators may modify dishes.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dish'
      responses:
        '200':
          description: Dish updated
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /dish/reset-cost:
    post:
      tags: [Dishes]
      summary: Reset dish price
      description: Only administrators may modify dish prices.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DishCostUpdate'
      responses:
        '200':
          description: Price updated
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /dish/add-ingredient:
    post:
      tags: [Dishes]
      summary: Attach an ingredient to a dish
      description: Only administrators may change composed dishes.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DishIngredientData'
      responses:
        '200':
          description: Ingredient added
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /dish/remove-ingredient:
    post:
      tags: [Dishes]
      summary: Detach an ingredient from a dish
      description: Only administrators may change composed dishes.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DishIngredientData'
      responses:
        '200':
          description: Ingredient removed
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /dish/delete/{id}:
    delete:
      tags: [Dishes]
      summary: Delete a dish
      description: Only administrators may delete dishes.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Dish deleted
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: The caller does not have the required role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        details:
          type: string
      additionalProperties: false
    AuthData:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
          description: Three-digit login code
      required: [id, code]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: Bearer token to send in the Authorization header
      required: [token]
    BillCreationRequest:
      type: object
      properties:
        tableNumber:
          $ref: '#/components/schemas/TableNumber'
        guestNumber:
          type: integer
          format: int32
          minimum: 1
        birthday:
          type: boolean
      required: [tableNumber, guestNumber]
    BillCreationResponse:
      type: object
      properties:
        billId:
          type: integer
          format: int64
        bonusPoints:
          type: integer
          format: int32
        message:
          type: string
      required: [billId, bonusPoints, message]
    BillLookupResponse:
      type: object
      properties:
        bill:
          $ref: '#/components/schemas/Bill'
        bonusPoints:
          type: integer
          format: int32
      required: [bill, bonusPoints]
    BillPaymentResponse:
      type: object
      properties:
        billId:
          type: integer
          format: int64
        message:
          type: string
      required: [billId, message]
    Bill:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        time:
          type: string
          format: date-time
        sum:
          type: number
          format: double
        billStatus:
          $ref: '#/components/schemas/BillStatus'
        guestNumber:
          type: integer
          format: int32
        bonusPoints:
          type: integer
          format: int32
          readOnly: true
      required: [id, time, sum, billStatus]
    DishWithIngredientsRequest:
      type: object
      properties:
        dish:
          $ref: '#/components/schemas/Dish'
        ingredients:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/DishIngredientAmount'
      required: [dish, ingredients]
    DishIngredientAmount:
      type: object
      properties:
        ingredientId:
          type: integer
          format: int64
        amount:
          type: number
          format: double
      required: [ingredientId, amount]
    DishIngredientData:
      allOf:
        - $ref: '#/components/schemas/DishIngredientAmount'
        - type: object
          properties:
            dishId:
              type: integer
              format: int64
          required: [dishId]
    DishCostUpdate:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cost:
          type: number
          format: double
      required: [id, cost]
    Employee:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        positions:
          $ref: '#/components/schemas/Positions'
        name:
          type: string
        secondName:
          type: string
        patronymic:
          type: string
          nullable: true
        birthDate:
          type: string
          format: date
        sex:
          type: string
          maxLength: 1
      required: [positions, name, secondName, birthDate, sex]
...
    Dish:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        name:
          type: string
        primeCost:
          type: number
          format: double
          readOnly: true
        cost:
          type: number
          format: double
        preparingTime:
          type: integer
          format: int32
        isKitchen:
          type: boolean
          description: True for kitchen dishes, false for bar items
      required: [name, cost, preparingTime, isKitchen]
    BillStatus:
      type: string
      enum: [open, paid]
    TableNumber:
      type: string
      enum: [T1, T2, T3, T4, T5, T6, T7, T8, T9, T10]
...
    Positions:
      type: string
      enum: [waiter, cook, barman, admin]